client/requirements.txt
PILLOW
Adafruit_GPIO